src/requirements.txt
django==2.2.7
psycopg2==2.8.4
django-pagedown==2.0.3
django-markdown-deux==1.0.5
